# filters/builtin: minor performance optimization for sanitize in StripQuery Filter