@AlexanderYastrebov 2024-02-16
Please include benchstat results (without 10+10 before/after benchmark outputs) into commit message (and update PR description).

I'd also suggest to prefix commit message with a package name, e.g.:
`filters/builtin: introduce benchmark for stripQuery filter`