@trkohler 2024-02-16
I attached benchmark results in last commit (https://github.com/zalando/skipper/pull/2940/commits/afdd4268f0b22934288e2d165f7e901f0f586b81)

I have also tried to use `AppendQuoteToASCII`, but it resulted in some failed tests. And when I do atempts to fix it properly, there are more allocations happening and it degrades benchmark result. Probably, I use it incorrectly... For now, I let the line be unchanged. 

I've also re-wrote description pointing out my motivation for a contribution. I'm sorry if previously it was thought as some issue with time-pressure. It is not pressuring at all, it's just my attempt to do an enhancement on some small part of source code. I thought that if it would succeed, I can try something bigger next.