@AlexanderYastrebov 2024-02-19
You'd need to `git commit --amend -s` to add DCO tothe last commit.
Please include benchstat results (**without** individual benchmark outputs) to see better see the improvement.